3-2プロジェクトでコンポーネント開発！/Rectコンポーネントを別ファイルに！
