+++ react_app/src/App.js
@@ -1,46 +1,19 @@
 import React, {Component} from 'react';//Reactオブジェクトの中のComponent
 //import logo from './logo.svg';
 import './App.css';
+import Rect from './Rect.js';
 
 class App extends Component {
   render(){
     return<div>
-      <Rect x="50" y="50" w="150" h="150" c="cyan"/>
-      <Rect x="150" y="100" w="150" h="150" c="magenta"/>
-      <Rect x="100" y="150" w="150" h="150" c="black"/>
+      <Rect x="50" y="50" w="150" h="150" c="#6ff9" r="50"/>
+      <Rect x="150" y="100" w="150" h="150" c="#f6f9"r="75"/>
+      <Rect x="100" y="150" w="150" h="150" c="#6669" r="25"/>
     </div>;
   }
 }
 
-class Rect extends Component{
-  x=0;
-  y=0;
-  width=0;
-  height=0;
-  color="white";
-  style={};
 
-  constructor(props){
-    super(props);
-    this.x=props.x;
-    this.y=props.y;
-    this.width=props.w;
-    this.height=props.h;
-    this.color=props.c;
-    this.style={
-      backgroundColor:this.color,
-      position:"absolute",
-      left:this.x+"px",
-      top:this.y+"px",
-      width:this.width+"px",
-      height:this.height+"px"
-    }
 
-  }
-
-  render(){
-    return <div style={this.style}></div>;
-  }
-}
 
 export default App;//class Appを使えるようにエクスポートしておく
